Amend script to adapt to schema version
Supports versions 0.2 and 0.3 and duplicates the
rdl-02 JKAN schema without amending it. As the new
schema hardens, we can make changes to rdl-03,
but for now it's a sufficient shorter path to get
the new rdl-03 datasets displaying in the same
way as the rdl-02 datasets in the UI. There are
some TODOs added for unresolved questions related
to summarizations and other discrepancies between
the v0.3 YAML schema and the official JSON schema.